Make updates better

+++ functions/_scuba_sub_install.fish
@@ -2,7 +2,7 @@ function _scuba_sub_install
     if not set -q argv[1]
         return
     end
-
+    argparse --stop-nonopt 'u-updating' -- $argv
     set -l arg $argv[1]
 
     printf '%s\n' "Installing $arg..."
@@ -12,26 +12,28 @@ function _scuba_sub_install
     set -l argEscaped (string escape --style=var $arg)
     set -l location /tmp/scuba/$argEscaped
 
-    rm -rf $location
+    if not set -q _flag_updating
+        rm -rf $location
 
-    if test -e $argSplit[1]
-        cp -r $argSplit[1] $location
-    else if GIT_TERMINAL_PROMPT=0 git clone https://github.com/$argSplit[1] $location
-    else
-        printf '%s' (set_color --bold red) "error: " (set_color normal) "target not found: $argSplit[1]" \n
-        exec fish --init-command="set -g fish_greeting; _scuba_sub_install $argv[2..-1]"
-    end
+        if test -e $argSplit[1]
+            cp -r $argSplit[1] $location
+        else if GIT_TERMINAL_PROMPT=0 git clone https://github.com/$argSplit[1] $location
+        else
+            printf '%s' (set_color --bold red) "error: " (set_color normal) "target not found: $argSplit[1]" \n
+            exec fish --init-command="set -g fish_greeting; _scuba_sub_install $argv[2..-1]"
+        end
 
-    if set -q argSplit[2] && not git -c advice.detachedHead=false -C $location checkout $argSplit[2]
-        printf '%s' (set_color --bold red) "error: " (set_color normal) "target not found: $argSplit[2]" \n
-        exec fish --init-command="set -g fish_greeting; _scuba_sub_install $argv[2..-1]"
+        if set -q argSplit[2] && not git -c advice.detachedHead=false -C $location checkout $argSplit[2]
+            printf '%s' (set_color --bold red) "error: " (set_color normal) "target not found: $argSplit[2]" \n
+            exec fish --init-command="set -g fish_greeting; _scuba_sub_install $argv[2..-1]"
+        end
+
+        if not contains $arg $_scuba_plugins
+            set -Ua _scuba_plugins $arg
+        end
     end
 
     set -U _scuba_"$argEscaped"_sha (git -C $location rev-parse HEAD 2>/dev/null)
-
-    if not contains $arg $_scuba_plugins
-        set -Ua _scuba_plugins $arg
-    end
 
     mkdir -p $location/{completions,conf.d,functions}
 
@@ -44,13 +46,18 @@ function _scuba_sub_install
     set -U _scuba_"$argEscaped"_files (string replace $location '' $location/{completions,conf.d,functions}/**)
     set -l fileVarName _scuba_"$argEscaped"_files
 
-    for file in (basename -s .fish $$fileVarName)
-        emit "$file"_install
-    end
-
     set_color --bold blue
     printf '%s\n' "$arg installed!"
     set_color normal
 
-    exec fish --init-command="set -g fish_greeting; _scuba_sub_install $argv[2..-1]"
+    exec fish --init-command="set -g fish_greeting
+    if test -n \"$_flag_updating\"
+        for file in (basename -s .fish $$fileVarName)
+            emit \$file'_update'
+        end
+    end
+    for file in (basename -s .fish $$fileVarName)
+        emit \$file'_install'
+    end
+    _scuba_sub_install $argv[2..-1]"
 end
